Release 1.0.0

+++ deblocking/gestures/view/HGestureView.m
@@ -161,7 +161,7 @@
     
     for (UIButton *btn in self.selectedBtnArray) {
         
-        [selectStr appendFormat:@"%ld",btn.tag];
+        [selectStr appendFormat:@"%ld",(long)btn.tag];
     }
     
     NSString *pwdStr = [[NSUserDefaults standardUserDefaults] objectForKey:@"pwd"];
